+++ src/modules/news/layout.js
@@ -1,0 +1,5 @@
+import AnnouncementBar from './components/AnnouncementBar.vue'
+
+export default {
+  'header:before': AnnouncementBar
+}